UPSTREAM: kbuild: remove cc-name variable
There is one more user of $(cc-name) in the top Makefile. It is supposed
to detect Clang before invoking Kconfig, so it should still be there
in the $(shell ...) form. All the other users of $(cc-name) have been
replaced with $(CONFIG_CC_IS_CLANG). Hence, scripts/Kbuild.include does
not need to define cc-name any more.

Change-Id: Ie9d2a1e142238f5821a6610581ca997a49004463
Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Signed-off-by: Santhosh <santhosh.user.why.red@gmail.com>